+++ CHANGELOG.md
@@ -28,6 +28,7 @@ Changes:
 * Clarified that all sources that come in from the Polyglot API `eval` method
   will be treated as UTF-8, and cannot be re-interpreted as another encoding
   using a magic comment.
+* `-Xembedded` can now set set on the launcher command line.
 
 # 1.0 RC 5, August 2018